% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.raster.r
\name{extend.raster}
\alias{extend.raster}
\title{Extend a BIL raster file by NOVALUE data}
\usage{
extend.raster(x, new.extent, write.raster = TRUE)
}
\arguments{
\item{x}{a raster object}

\item{new.extent}{either an extent or a numerical vector that can be coerced to an extent.}

\item{write.raster}{logical. Should the extended raster be written to disk (TRUE) or returned (FALSE)?}
}
\value{
a raster object if \code{write.raster == FALSE}
}
\description{
Extend an existing raster by NOVALUE data and opionally write it as a BIL raster file
}
\seealso{
\link{readINCABIL}
   \link{writeINCABIL}
}
\author{
Simon Frey
}
