% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.COSERO.input.r
\name{make.COSERO.input}
\alias{make.COSERO.input}
\title{Create a T_IZ or P_IZ file from a set of INCA-Files}
\usage{
make.COSERO.input(f, shape, nzraster, output = NULL, otf = FALSE,
  sortbynz = TRUE, ...)
}
\arguments{
\item{f}{A character vector of INCA files}

\item{shape}{Absolute path to a shapefile (points) at which locations the values of the INCA files are extracted}

\item{nzraster}{Path to a raster with the nz information}

\item{output}{name of the outputfile or NULL}

\item{otf}{write on-the-fly (don't hold the whole matrix in memory)}

\item{...}{parameters passed to readINCABIL}
}
\value{
if output == NULL the IZ-matrix is returned. Else the IZ-File is written and nothing is returned to R.
}
\description{
Create a T_IZ or P_IZ file from a set of INCA-Files
}
\details{
Reads a set of INCA files (Precipitation or Temperature) and calculates a matrix that can be read by COSERO.

    The shapefile must be projected in the same coordinate systm as the nzraster as well as the INCA files.
    Normally this is "+proj=lcc +lat_1=46 +lat_2=49 +lat_0=47.5 +lon_0=13.33333333333333 +x_0=400000 +y_0=400000 +ellps=bessel +units=m +no_defs"
}
\seealso{
\link{readINCABIL}
}
\author{
Simon Frey
}
