% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.hdr.r
\name{read.hdr}
\alias{read.hdr}
\title{Read an HDR file}
\usage{
read.hdr(file, add.nblocks = TRUE)
}
\arguments{
\item{file}{character string. Filename of the file to be read.}

\item{add.nblocks}{logical. Should the agrument nblocks be added to the output? See details.}
}
\value{
a data.frame containing the arguments of the hdr file
}
\description{
Read an HDR file which is the header part of a BIL raster file
}
\details{
If add.nblocks is TRUE (the default) then the argument \code{nblocks} is added to the output list (if it does nikt already exist).
    If so, the value from the argument \code{nbands} is copied to \code{nblocks} and \code{nbands} is set to 1. This is because of handling
    in Delft-FEWS. See the wiki of Delft-FEWS: (https://publicwiki.deltares.nl/display/FEWSDOC/BIL+BIP+BSQ+Parser).
}
\author{
Simon Frey
}
