% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.spatacus.r
\name{read.spartacus}
\alias{read.spartacus}
\title{read.spartacus}
\usage{
read.spartacus(x, crs = "default")
}
\arguments{
\item{x}{either a single rda file or a vector of rda files containing spartacus data}

\item{crs}{a crs chracter string providing the spatial reference information of the spartacus file}
}
\value{
a raster brick
}
\description{
Read information from ZAMG's spatacus dataset
}
\author{
Simon Frey
}
