% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readZAMGGRIB.r
\name{readZAMGGRIB}
\alias{readZAMGGRIB}
\title{Read a GRIB-1 file using wgrib}
\usage{
readZAMGGRIB(
  gribfile,
  ex = NULL,
  crs = NULL,
  wgrib = NULL,
  recnr = NULL,
  variable = NULL,
  remove = FALSE
)
}
\arguments{
\item{gribfile}{Absolute path to a gribfile}

\item{ex}{Either an \code{\link{extent}} object, NULL, 'alaro', or 'arome'. As of now, 'arome' is not supported, yet.}

\item{crs}{Either a \code{\link{crs}} string, NULL, 'alaro', or 'arome'. As of now, 'arome' is not supported, yet.}

\item{wgrib}{Absolute path to the executable of wgrib. If NULL, wgrib must be installed properly on the system.}

\item{recnr}{numeric giving the record number of the respective variable within GRIB file}

\item{variable}{character string giving the abbreviated variable within the GRIB file.}

\item{remove}{either TRUE/FALSE for removing / not removing the gribfile after reading
or ''success'/fail' for only removing files that have / haven't been successfully read.}
}
\value{
A (projected) raster with the extracted information
}
\description{
Read a GRIB-1 file provided by the ZAMG using wgrib and extract a specific raster set.
}
\details{
If ex and/or crs are supplied, the returned raster will be projected. If ex and/or crs are specified as
"alaro" or "arome", the respective extent and crs are used. Note, that as of now, only alaro is supported.

    Either recnr or variable must be specified to extract the information of interest.
    If both are given the latter is ignored.

    This function relies on the package TigR, which is available on \href{https://github.com/freysimon/TigR}{Github}

    Note that wgrib must be installed on the machine. It can be downloaded from \url{http://www.cpc.ncep.noaa.gov/products/wesley/wgrib.html}
}
\seealso{
For reading INCA files, see \code{\link{readINCABIL}}
}
\author{
Simon Frey
}
