\name{substrRight}
\alias{substrRight}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Return the last n characters of a string.
}
\description{
}
\usage{
substrRight(x, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
a character vector.
}
\item{n}{
integer. Number of characters that are returned.
}
}
\details{
substrRight takes a character string and returns the last n characters.
}
\value{
a character string.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Simon Frey
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\link{substr}
}
\examples{
substrRight("abcdef", 2)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
