% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeINCABIL.r
\name{writeINCABIL}
\alias{writeINCABIL}
\title{Write a raster using the BIL format}
\usage{
writeINCABIL(x, file, delete.stx = TRUE, overwrite = TRUE, hdr = "keep")
}
\arguments{
\item{x}{raster object. The raster that will be written.}

\item{file}{character string. The file that will be written. Note that besides the .bil file, a .hdr and a .stx file are written, too.}

\item{delete.stx}{logical. Should the stx file be deleted?}

\item{overwrite}{logical. Should an existing file be overwritten?}

\item{hdr}{character string. Any of "keep", "delete", or "update".}
}
\description{
Write a raster in the BIL format used e.g. by INCA
}
\details{
\code{file} must be the filename of a raster file thar will be written. It may or may not have the file extension .bil. If not, it will be added to the srtring.

     The \code{stx}-file is not necessary for further usage of the raster file and may be delted using \code{delete.stx == TRUE}.

     The \code{hdr}-file matches the official ESRI specifications. However, Delft-FEWS needs the parameter \code{NBLOCKS} for reading time series. If \code{hdr == 'update'}
     the parameter \code{NBANDS} will be set to 1 and the new added parameter \code{NBLOCKS} will receive the former value of \code{NBANDS}.

     \code{hdr == 'delete'} deltes the \code{hdr}-file whereas \code{hdr == 'keep'} keeps it unedited.
}
\seealso{
\link{readINCABIL}
   \link{readZAMGGRIB}
   \link{read.hdr}
}
\author{
Simon Frey
}
