% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.raster.r
\name{extend.raster}
\alias{extend.raster}
\title{Extend a BIL raster file by NOVALUE data}
\usage{
extend.raster(
  x,
  new.extent,
  write.raster = TRUE,
  filename = NULL,
  mp = 1,
  noValue = NULL
)
}
\arguments{
\item{x}{a raster object. May be a raster, a RasterBrick or a RasterStack.}

\item{new.extent}{either an extent or a numerical vector that can be coerced to an extent.}

\item{write.raster}{logical. Should the extended raster be written to disk (TRUE) or returned (FALSE)?}

\item{filename}{character string. Filename used to write the extended raster. Only used if write.raster==TRUE.}

\item{mp}{numeric. multiplicator to multiply the (existing) raster with.}

\item{noValue}{numeric. Sets the nodata value for the new raster.}
}
\value{
a raster object if \code{write.raster == FALSE}
}
\description{
Extend an existing raster by NOVALUE data and opionally write it as a BIL raster file
}
\seealso{
\link{readINCABIL}
   \link{writeINCABIL}
}
\author{
Simon Frey
}
