% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.COSERO.input.r
\name{make.COSERO.input}
\alias{make.COSERO.input}
\title{Create a T_IZ or P_IZ file from a set of INCA-Files}
\usage{
make.COSERO.input(f, shape, nzraster, output = NULL, ...)
}
\arguments{
\item{f}{A character vector of INCA files}

\item{shape}{Absolute path to a shapefile (points) at which locations the values of the INCA files are extracted}

\item{nzraster}{Path to a raster with the nz information}

\item{output}{name of the outputfile or NULL}

\item{...}{parameters passed to readINCABIL}
}
\value{
if output == NULL the IZ-matrix is returned. Else the IZ-File is written and nothing is returned to R.
}
\description{
Create a T_IZ or P_IZ file from a set of INCA-Files
}
\seealso{
\link{readINCABIL}
}
\author{
Simon Frey
}
