% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readASC.r
\name{readASC}
\alias{readASC}
\title{read an ascii-raster-file}
\usage{
readASC(x, CoS = NULL)
}
\arguments{
\item{x}{character string pointing to a asc file}

\item{Cos}{(optional) character string in the format provided by \code{\link{proj4string}}}
}
\value{
a raster object
}
\description{
reads an asc raster file and returns the resulting raster
}
\details{
If CoS is provided, the resulting raster is projected using the given coordinate system. If not, an unprojected raster is returned.
}
\author{
Simon Frey
}
