% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readINCABIL.r
\name{readINCABIL}
\alias{readINCABIL}
\title{Read a binary INCA file}
\usage{
readINCABIL(filename, times = "first", date = NULL, remove = TRUE,
  CoSys = NULL, tz = "utc")
}
\arguments{
\item{filename}{A character string to the file to be read in}

\item{times}{A character string giving the elemnt of the BIL file to be read in. One of 'first', 'last', 'all', 'date'. See details.}

\item{date}{A POSIXct date or NULL. Only evaluated if times == date}

\item{remove}{Logical. Remove the processed files?}

\item{CoSys}{A character string giving the coordiante system of the raster or NULL.}

\item{tz}{A character string giving the time zone}
}
\value{
A named (date and time) list with rasters
}
\description{
Read an INCA file in the BIL format from a .tar.gz archive.
}
\details{
Both uncompressed and compressed (.tar.gz) files can be processed. If uncompressed, only the filename without extention must be provided.

   If times is "date", a POSIXct date object must be provided to selsect the date within the tim file.
   If times is "all", all timestamps within the file are read and a stack of rasters is retured.

   If remove = TRUE (the standard), the (uncompressed) files are deleted after they are processed. Does not affect compressed files!

   If CoSys is provided (e.g. from a \link{proj4string} command), the retured raster is projected using this reference system. Otherwise no coordinate system is used.
}
\examples{
   # file to load
   file <- paste(path.package("ZAMGR"),"/extdata/INCA_TT.tar.gz",sep="")

   # reading a compressed file
   x <- readINCABIL(file)

   # plot the results
   plot(x[[1]])

   # compress first and then read
   file <- untar(file, exdir = tempdir())
   file <- dir(path = tempdir(), pattern = ".bil", full.names = TRUE)
   x <- readINCABIL(file = gsub(".bil", replacement = "", file), times = "all")
}
\seealso{
\link{raster}
   \link{readBin}
   \link{proj4string}
   \link{crs}
}
\author{
Simon Frey
}
